sheep: upgrade 'idx' in object_cache from 32bit to 64bit
Change 'idx' in object cache from uint32_t to uint64_t for working correctly
on hyper volume.

+++ include/sheepdog_proto.h
@@ -419,7 +419,7 @@ static inline uint64_t vid_to_vdi_oid(uint32_t vid)
 	return VDI_BIT | ((uint64_t)vid << VDI_SPACE_SHIFT);
 }
 
-static inline uint64_t vid_to_data_oid(uint32_t vid, uint32_t idx)
+static inline uint64_t vid_to_data_oid(uint32_t vid, uint64_t idx)
 {
 	return ((uint64_t)vid << VDI_SPACE_SHIFT) | idx;
 }
